Implement Observer Design Pattern example
Added the `Article` class to represent articles with properties for `Title`, `Content`, and `PublishedAt`. Introduced the `Author` class as the subject, implementing the `ISubject<Article>` interface to manage followers and notify them of new articles. Added the `Follower` class as the observer, implementing the `IObserver<Article>` interface to handle notifications.

Defined the `IObserver<T>` and `ISubject<T>` interfaces to establish the Observer Design Pattern contract. Created a `.csproj` file targeting .NET 9.0 with nullable reference types and implicit usings enabled.

Added `Program.cs` to demonstrate the pattern with an author publishing articles and followers subscribing/unsubscribing and receiving notifications.